# openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/reports/":{"post":{"summary":"Create Report","description":"Creates a new incident report in the database, handling form data and file uploads.","operationId":"create_report_reports__post","requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_create_report_reports__post"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IncidentReport"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"summary":"List Reports","description":"Retrieves a list of incident reports, with optional filtering and pagination.","operationId":"list_reports_reports__get","parameters":[{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter reports by status (e.g., 'new', 'in_progress', 'closed')","title":"Status"},"description":"Filter reports by status (e.g., 'new', 'in_progress', 'closed')"},{"name":"start_date","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"description":"Filter reports created on or after this date (YYYY-MM-DD)","title":"Start Date"},"description":"Filter reports created on or after this date (YYYY-MM-DD)"},{"name":"end_date","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"description":"Filter reports created on or before this date (YYYY-MM-DD)","title":"End Date"},"description":"Filter reports created on or before this date (YYYY-MM-DD)"},{"name":"country","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter reports by country","title":"Country"},"description":"Filter reports by country"},{"name":"skip","in":"query","required":false,"schema":{"type":"integer","minimum":0,"description":"Number of reports to skip for pagination","default":0,"title":"Skip"},"description":"Number of reports to skip for pagination"},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"description":"Maximum number of reports to return","default":20,"title":"Limit"},"description":"Maximum number of reports to return"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/IncidentReport"},"title":"Response List Reports Reports  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/reports/analytics":{"get":{"summary":"Reports Analytics","description":"Provides analytics on incident reports, specifically counting occurrences of violation types.","operationId":"reports_analytics_reports_analytics_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Response Reports Analytics Reports Analytics Get"}}}}}}},"/reports/{report_id}":{"get":{"summary":"Get Report By Id","description":"Retrieves a single incident report by its unique report ID.","operationId":"get_report_by_id_reports__report_id__get","parameters":[{"name":"report_id","in":"path","required":true,"schema":{"type":"string","title":"Report Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IncidentReport"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"summary":"Update Report Status","description":"Updates the status of an existing incident report.","operationId":"update_report_status_reports__report_id__patch","parameters":[{"name":"report_id","in":"path","required":true,"schema":{"type":"string","title":"Report Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReportStatusUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IncidentReport"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_create_report_reports__post":{"properties":{"reporter_type":{"type":"string","title":"Reporter Type"},"anonymous":{"type":"boolean","title":"Anonymous","default":false},"email":{"anyOf":[{"type":"string","format":"email"},{"type":"null"}],"title":"Email"},"phone":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone"},"preferred_contact":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Preferred Contact"},"date":{"type":"string","format":"date-time","title":"Date","description":"Date of the incident (YYYY-MM-DD or YYYY-MM-DDTHH:MM:SS)"},"country":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Country"},"city":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"City"},"longitude":{"type":"number","maximum":180.0,"minimum":-180.0,"title":"Longitude","description":"Longitude of the incident location"},"latitude":{"type":"number","maximum":90.0,"minimum":-90.0,"title":"Latitude","description":"Latitude of the incident location"},"description":{"type":"string","minLength":3,"title":"Description","description":"Detailed description of the incident"},"violation_types":{"type":"string","title":"Violation Types","description":"Comma-separated list of violation types (e.g., 'detention,torture')"},"files":{"anyOf":[{"items":{"type":"string","format":"binary"},"type":"array"},{"type":"null"}],"title":"Files"}},"type":"object","required":["reporter_type","date","longitude","latitude","description","violation_types"],"title":"Body_create_report_reports__post"},"ContactInfo":{"properties":{"email":{"anyOf":[{"type":"string","format":"email"},{"type":"null"}],"title":"Email"},"phone":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone"},"preferred_contact":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Preferred Contact"}},"type":"object","title":"ContactInfo"},"Coordinates":{"properties":{"type":{"type":"string","title":"Type","default":"Point"},"coordinates":{"items":{"type":"number"},"type":"array","title":"Coordinates"}},"type":"object","required":["coordinates"],"title":"Coordinates"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"IncidentDetails":{"properties":{"date":{"type":"string","format":"date-time","title":"Date"},"location":{"$ref":"#/components/schemas/Location"},"description":{"type":"string","title":"Description"},"violation_types":{"items":{"type":"string"},"type":"array","title":"Violation Types"}},"type":"object","required":["date","location","description","violation_types"],"title":"IncidentDetails"},"IncidentReport":{"properties":{"_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Id"},"report_id":{"type":"string","title":"Report Id"},"reporter_type":{"type":"string","title":"Reporter Type"},"anonymous":{"type":"boolean","title":"Anonymous","default":false},"contact_info":{"anyOf":[{"$ref":"#/components/schemas/ContactInfo"},{"type":"null"}]},"incident_details":{"$ref":"#/components/schemas/IncidentDetails"},"evidence":{"items":{"$ref":"#/components/schemas/ReportEvidence"},"type":"array","title":"Evidence","default":[]},"status":{"type":"string","title":"Status","default":"new"},"assigned_to":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Assigned To"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["report_id","reporter_type","incident_details"],"title":"IncidentReport","example":{"anonymous":false,"contact_info":{"email":"reporter@example.com","phone":"123-456-7890","preferred_contact":"email"},"created_at":"2025-02-08T12:00:00.000Z","evidence":[{"description":"Photo of the incident location","type":"photo","url":"/uploads/IR-2025-1001_image1.jpg"}],"incident_details":{"date":"2025-06-01T00:00:00Z","description":"Witnessed arbitrary detention.","location":{"city":"Ramallah","coordinates":{"coordinates":[35.2163,31.9037],"type":"Point"},"country":"Palestine"},"violation_types":["arbitrary detention","freedom of movement"]},"report_id":"IR-2025-1001","reporter_type":"individual","status":"new"}},"Location":{"properties":{"country":{"type":"string","title":"Country"},"city":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"City"},"coordinates":{"$ref":"#/components/schemas/Coordinates"}},"type":"object","required":["country","coordinates"],"title":"Location"},"ReportEvidence":{"properties":{"type":{"type":"string","title":"Type"},"url":{"type":"string","title":"Url"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"}},"type":"object","required":["type","url"],"title":"ReportEvidence"},"ReportStatusUpdate":{"properties":{"status":{"type":"string","pattern":"^(new|in_progress|closed|resolved|on_hold)$","title":"Status","description":"Status of the report"}},"type":"object","required":["status"],"title":"ReportStatusUpdate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}